Move create invites service to invites folder
