+++ app/services/tasks/update.rb
@@ -27,7 +27,7 @@ module Tasks
       if type == "Tasks::Watchlist"
         user.invites.where(talent_invite: false).update_all(max_uses: nil)
       elsif type == "Tasks::ShareProfile" && user.talent
-        service = CreateInvite.new(user_id: user.id, single_use: true, talent_invite: true)
+        service = Invites::Create.new(user_id: user.id, single_use: true, talent_invite: true)
         service.call
       elsif type == "Tasks::Register"
         Reward.create!(user: user, amount: 100, category: "quest", reason: "Got 5 people to register")
