+++ lib/tasks/dev.rake
@@ -15,8 +15,8 @@ if Rails.env.development?
       Feed.create!(user: admin)
       Tasks::PopulateForUser.new.call(user: admin)
 
-      investor_invite = CreateInvite.new(user_id: admin.id).call
-      talent_invite = CreateInvite.new(user_id: admin.id, talent_invite: true).call
+      investor_invite = Invites::Create.new(user_id: admin.id).call
+      talent_invite = Invites::Create.new(user_id: admin.id, talent_invite: true).call
 
       investor_result = CreateUser.new.call(
         email: "investor@talentprotocol.com",
